restructure faucet explorer
