--- src/icons/mod.rs
@@ -1,5 +1,0 @@
-mod check_icon;
-mod lightning_icon;
-
-pub use check_icon::CheckIcon;
-pub use lightning_icon::LightningIcon;
