+++ src/utils/mod.rs
@@ -1,0 +1,6 @@
+pub mod address;
+pub mod error;
+pub mod format;
+pub mod icons;
+pub mod key;
+pub mod message;
